+++ example-service/main.tf
@@ -17,18 +17,7 @@ terraform {
     }
   }
 
-  backend "s3" {
-    region         = "us-east-2"
-    dynamodb_table = "terraform"
-    encrypt        = "true"
-
-    # must be unique to your AWS account; try replacing
-    # uiuc-tech-services-sandbox with the friendly name of your account
-    bucket = "terraform.uiuc-tech-services-sandbox.aws.illinois.edu" #FIXME
-
-    # must be unique (within bucket) to this repository + environment
-    key = "FIXME/example-service/terraform.tfstate" #FIXME
-  }
+  # see backend.tf for remote state configuration
 }
 
 ## Inputs (specified in terraform.tfvars)