+++ vpc/endpoints.tf
@@ -1,46 +1,52 @@
-# Example configuration for VPC Endpoints
+# Configuration for VPC Endpoints
 # https://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/vpc-endpoints.html
 #
 # Copyright (c) 2018 Board of Trustees University of Illinois
 
+## Inputs (specified in terraform.tfvars)
+
+# Gateway VPC Endpoints, created by default since there is no additional charge
+# https://docs.aws.amazon.com/vpc/latest/privatelink/vpce-gateway.html#gateway-endpoint-pricing
+variable "gateway_vpc_endpoint_service_names" {
+  description = "List of ServiceNames with region replaced by '{{REGION}}'"
+  type        = list(string)
+  default     = [
+    "com.amazonaws.{{REGION}}.dynamodb",
+    "com.amazonaws.{{REGION}}.s3",
+  ]
+}
+
+# Interface VPC Endpoints for AWS services, not created by default since each
+# interface incurs cost and consumes an IP address.  Specify the ones you need.
+#
 # Hint: use `aws ec2 describe-vpc-endpoint-services` to find more service names
+variable "interface_vpc_endpoint_service_names" {
+  description = "List of ServiceNames with region replaced by '{{REGION}}'"
+  type        = list(string)
+  default     = []
+}
+
+variable "interface_vpc_endpoint_subnets" {
+  description = "Specify one subnet per Availability Zone to be used for Interface VPC Endpoints (may be public-, campus-, or private-facing).  Maps availability zone suffix (e.g. 'a' for us-east-2a) to subnet key"
+  type        = map(string)
+  default     = {}
+}
+
+## Resources
 
 locals {
-  # Gateway VPC Endpoints, enabled by default
-  gateway_vpc_endpoint_service_names = [
-    "com.amazonaws.${var.region}.dynamodb",
-    "com.amazonaws.${var.region}.s3",
-  ]
-
-  # Interface VPC Endpoints, disabled by default because they create elastic
-  # network interfaces (thus consuming an IP) on each chosen subnet.  Uncomment
-  # the ones you want to use.
-  interface_vpc_endpoint_service_names = [
-    #"com.amazonaws.${var.region}.ec2",
-    #"com.amazonaws.${var.region}.ec2messages",
-    #"com.amazonaws.${var.region}.elasticloadbalancing",
-    #"com.amazonaws.${var.region}.kinesis-streams",
-    #"com.amazonaws.${var.region}.kms",
-    #"com.amazonaws.${var.region}.servicecatalog",
-    #"com.amazonaws.${var.region}.sns",
-    #"com.amazonaws.${var.region}.ssm",
-  ]
-
   # which subnets from main.tf to use for Interface VPC Endpoints
-  interface_vpc_endpoint_subnet_ids = [module.private1-a-net.id, module.private1-b-net.id]
-
-  # derived values used in main.tf
-  gateway_vpc_endpoint_ids = {
-    for k in local.gateway_vpc_endpoint_service_names :
-    # conditional is needed for clean `terraform destroy` (as of TF 0.12.9)
-    k => (contains(keys(aws_vpc_endpoint.gateway), k) ? aws_vpc_endpoint.gateway[k].id : "")
-  }
+  interface_vpc_endpoint_subnet_ids = [for k,v in var.interface_vpc_endpoint_subnets :
+    try(module.public-facing-subnet[v].id, module.campus-facing-subnet[v].id, module.private-facing-subnet[v].id,
+      # https://github.com/hashicorp/terraform/issues/15469#issuecomment-515240849
+      # so we know which occurrence failed
+      file("\nERROR: var.interface_vpc_endpoint_subnets contains unexpected subnet '${v}'"))]
 }
 
 # create Gateway VPC Endpoints (if desired)
 
 resource "aws_vpc_endpoint" "gateway" {
-  for_each = toset(local.gateway_vpc_endpoint_service_names)
+  for_each = toset([for x in var.gateway_vpc_endpoint_service_names : replace(x, "{{REGION}}", var.region)])
 
   tags = merge(var.tags, {
     Name = "${var.vpc_short_name}-vpce-${each.value}"
@@ -54,7 +60,7 @@ resource "aws_vpc_endpoint" "gateway" {
 # create Interface VPC Endpoints (if desired)
 
 resource "aws_vpc_endpoint" "interface" {
-  for_each = toset(local.interface_vpc_endpoint_service_names)
+  for_each = toset([for x in var.interface_vpc_endpoint_service_names : replace(x, "{{REGION}}", var.region)])
 
   tags = merge(var.tags, {
     Name = "${var.vpc_short_name}-vpce-${each.value}"
@@ -74,7 +80,7 @@ resource "aws_vpc_endpoint" "interface" {
 # but our SG permits IPv6 anyway in case that changes
 
 resource "aws_security_group" "endpoints" {
-  count = length(local.interface_vpc_endpoint_service_names) > 0 ? 1 : 0
+  count = length(var.interface_vpc_endpoint_service_names) > 0 ? 1 : 0
 
   tags = merge(var.tags, {
     Name = "${var.vpc_short_name}-vpc-endpoints"
@@ -87,7 +93,7 @@ resource "aws_security_group" "endpoints" {
 # allow all outbound
 resource "aws_security_group_rule" "endpoint_egress" {
   # note: tags not supported
-  count = length(local.interface_vpc_endpoint_service_names) > 0 ? 1 : 0
+  count = length(var.interface_vpc_endpoint_service_names) > 0 ? 1 : 0
 
   security_group_id = aws_security_group.endpoints[0].id
   type              = "egress"
@@ -101,7 +107,7 @@ resource "aws_security_group_rule" "endpoint_egress" {
 # allow inbound only from this VPC
 resource "aws_security_group_rule" "endpoint_ingress" {
   # note: tags not supported
-  count = length(local.interface_vpc_endpoint_service_names) > 0 ? 1 : 0
+  count = length(var.interface_vpc_endpoint_service_names) > 0 ? 1 : 0
 
   security_group_id = aws_security_group.endpoints[0].id
   type              = "ingress"